renamed store
allFactionsArrayStore has been renamed to more generic, as we also insantiate another store from the same file
