+++ src/component/SvelteStore.js
